+++ src/frontend/magnify/packages/components/src/utils/configuration.ts
@@ -1,3 +1,5 @@
+import { Maybe } from '../types/misc';
+
 export {};
 declare global {
   interface Window {
@@ -6,6 +8,7 @@ declare global {
       KEYCLOAK_CLIENT_ID: string;
       KEYCLOAK_URL: string;
       KEYCLOAK_EXPIRATION_SECONDS: number;
+      LANGUAGE_CODE: Maybe<string>;
       KEYCLOAK_REALM: string;
       JITSI_DOMAIN: string;
     };